==> code/notebooks/0_first_try.ipynb <==
import os
import json
import pandas as pd

df = pd.read_json('../data/raw/yelp_academic_dataset_review.json', lines=True)

df.head()

business_data = pd.read_json('../data/raw/yelp_academic_dataset_business.json', lines=True)

business_data.columns

business_data['attributes'].head()

df = pd.merge(df, business_data, on='business_id', how='left')

df.head()

df_ihop = df[df['name'] == 'IHOP']

df_ihop.head()

df_ihop[df_ihop].shape

df_ihop.groupby('state').size().reset_index(name='conteo')

df_ihop[df_ihop['state']=='FL'].shape

# Ensure 'date' is in datetime format
df_ihop.loc[:, 'date'] = pd.to_datetime(df_ihop['date'])

# Filter for rows where state is 'FL', group by month, and count
florida_counts = df_ihop[df_ihop['state'] == 'FL'].groupby(df_ihop['date'].dt.to_period('M')).size().reset_index(name='count')

# Rename the columns for clarity
florida_counts.rename(columns={'date': 'month'}, inplace=True)

florida_counts['count'].max()

df_ihop[(df_ihop['date'].dt.year >= 2015) & 
                      (df_ihop['state'].isin(['FL', 'PA', 'LA']))].shape

df[(df['name'] == 'IHOP') & (df['date'].dt.year >= 2015) & 
                      (df['state'].isin(['FL', 'PA', 'LA']))]

df_ihop.columns

attributes_expanded = pd.json_normalize(business_data['attributes'])

# Merge expanded attributes back into the original DataFrame
business_data = pd.concat([business_data.drop(['attributes'], axis=1), attributes_expanded], axis=1)

business_data.head()

business_data.columns

df_ihop_attributes = business_data[business_data['name']=='IHOP'].copy()

selected_columns = [
    'BikeParking', 'RestaurantsDelivery',
    'WiFi', 'BusinessParking', 'HasTV',
    'NoiseLevel'
]

# Summary statistics for the selected columns
summary_stats = df_ihop_attributes[selected_columns].describe(include='all')
summary_stats
